refactor panda loader related functions to panda_loader from df_loaders
